docs: Claude Code Guidelines 추가 및 컨트롤러 구조 개선
- Claude Code Guidelines by Sabrina Ramonov 추가
- 코딩 베스트 프랙티스 및 단축키 정의
- BaseController 제거 및 중복 코드 각 컨트롤러에 통합
- TDD, 함수 품질, 테스트 작성 가이드라인 포함
- 일관된 예외 처리 및 도메인 어휘 사용 원칙 적용
